Refactored TraceV3 to stand-alone file and added tests
