+++ scripts/UnifiedLogReader.py
@@ -48,6 +48,7 @@ import time
 from UnifiedLog import __version__
 from UnifiedLog import Lib as UnifiedLogLib
 from UnifiedLog import logger
+from UnifiedLog import tracev3_file
 from UnifiedLog import virtual_file
 from UnifiedLog import virtual_file_system
 
@@ -187,7 +188,8 @@ def RecurseProcessLogFiles(input_path, ts_list, uuidtext_folder_path, caches, pr
         if file_name.lower().endswith('.tracev3') and not file_name.startswith('._'):
             logger.info("Trying to read file - " + input_file_path)
             file_object = virtual_file.VirtualFile(input_file_path, 'traceV3')
-            UnifiedLogLib.TraceV3(vfs, file_object, ts_list, uuidtext_folder_path, caches).Parse(proc_func)
+            trace_file = tracev3_file.TraceV3(vfs, file_object, ts_list, uuidtext_folder_path, caches)
+            trace_file.Parse(proc_func)
         elif os.path.isdir(input_file_path):
             RecurseProcessLogFiles(input_file_path, ts_list, uuidtext_folder_path, caches, proc_func)
 
@@ -312,7 +314,8 @@ def main():
             RecurseProcessLogFiles(tracev3_path, ts_list, uuidtext_folder_path, caches, proc_func)
         else:
             file_object = virtual_file.VirtualFile(input_file_path, 'traceV3')
-            UnifiedLogLib.TraceV3(vfs, file_object, ts_list, uuidtext_folder_path, caches).Parse(proc_func)
+            trace_file = tracev3_file.TraceV3(vfs, file_object, ts_list, uuidtext_folder_path, caches)
+            trace_file.Parse(proc_func)
         if f:
             f.close()
         if db_conn:
