+++ tests/virtual_file.py
@@ -17,13 +17,13 @@ class VirtualFileTest(test_lib.BaseTestCase):
     def testGetFileSize(self):
         '''Tests the get_file_size function.'''
         path = self._GetTestFilePath(['0D3C2953A33917B333DD8366AC25F2'])
-        file_object = virtual_file.VirtualFile(path, filetype='uuidtext')
+        file_entry = virtual_file.VirtualFile(path, filetype='uuidtext')
 
-        file_object.open()
+        file_entry.open()
         try:
-            file_size = file_object.get_file_size()
+            file_size = file_entry.get_file_size()
         finally:
-             file_object.close()
+            file_entry.close()
 
         self.assertEqual(file_size, 33)
