+++ tests/virtual_file_system.py
@@ -36,11 +36,12 @@ class VirtualFileSystemTests(test_lib.BaseTestCase):
             virtual_file.VirtualFile)
 
         expected_directory_entries = [
+            '0000000000000030.tracev3',
             '0D3C2953A33917B333DD8366AC25F2',
             '8E21CAB1DCF936B49F85CF860E6F34EC']
 
         directory_entries = file_system.listdir(self._TEST_DATA_PATH)
-        self.assertEqual(len(directory_entries), 2)
+        self.assertEqual(len(directory_entries), 3)
         self.assertEqual(sorted(directory_entries), expected_directory_entries)
 
     def testIsDir(self):
@@ -74,9 +75,9 @@ class VirtualFileSystemTests(test_lib.BaseTestCase):
 
         path = os.path.join(
             self._TEST_DATA_PATH, '0D3C2953A33917B333DD8366AC25F2')
-        file_object = file_system.get_virtual_file(path, filetype='uuidtext')
-        self.assertIsNotNone(file_object)
-        self.assertIsInstance(file_object, virtual_file.VirtualFile)
+        file_entry = file_system.get_virtual_file(path, filetype='uuidtext')
+        self.assertIsNotNone(file_entry)
+        self.assertIsInstance(file_entry, virtual_file.VirtualFile)
 
 
 if __name__ == '__main__':
